PF int a long

+++ src/main/java/com/guardias/backend/entity/Adicional.java
@@ -10,22 +10,22 @@ public class Adicional {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-    private int id;
+    private long id;
     private String nombre;
 
     public Adicional() {
     }
 
-    public Adicional(int id, String nombre) {
+    public Adicional(long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
