+++ src/main/java/com/guardias/backend/entity/Categoria.java
@@ -11,22 +11,22 @@ public class Categoria {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-    private int id;
+    private long id;
     private String nombre;
 
     public Categoria() {
     }
 
-    public Categoria(int id, String nombre) {
+    public Categoria(long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
