+++ src/main/java/com/guardias/backend/entity/Region.java
@@ -10,22 +10,22 @@ public class Region {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String nombre;
 
     public Region() {
     }
 
-    public Region(int id, String nombre) {
+    public Region(long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
